hospital-ai-receptionist/api-server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
gspread>=6.1.0
google-auth>=2.29.0
apscheduler>=3.10.0
sendgrid>=6.11.0
httpx>=0.27.0
huggingface-hub>=0.23.0
python-dateutil>=2.9.0
